Move connected-component helpers into tools
